+++ app/__init__.py
@@ -2,7 +2,7 @@ import os
 from flask import Flask
 from .socketio_ext import socketio
 
-VERSION = '1.9.6'
+VERSION = '1.9.7'
 
 
 def create_app():